(Partial) fix for bug around RVM detection.
When running multiple Jenkins build agents (slaves), particularly if
these are created and destroyed independently of the master instance,
the caching of the `rvm_path` on the master prevents the absence of
on `rvm` installation on the current slave being detected.

Note that the value is still being held globally, so two concurrently
executing jobs still have the potential to fail to detect the presence
of `rvm`, though the likelihood is deminished.